+++ packages/ui/src/components/molecules/Notifications/generateRandomID.ts
@@ -1,0 +1,6 @@
+export function generateRandomID() {
+  const crypto = window.crypto || window.Crypto
+  const array = new Uint32Array(1)
+  const id = crypto.getRandomValues(array).toString()
+  return id
+}